--- frontend/src/employee-frontend/api/finance-basics.ts
@@ -1,5 +1,0 @@
-// SPDX-FileCopyrightText: 2017-2022 City of Espoo
-//
-// SPDX-License-Identifier: LGPL-2.1-or-later
-
-export type FeeThresholdsSaveError = 'date-overlap'
